+++ sickrage/core/helpers/encoding.py
@@ -70,10 +70,12 @@ def ek(f):
         else:
             result = f(*[ss(x) if isinstance(x, six.string_types) else x for x in args], **kwargs)
 
+        if isinstance(result, types.GeneratorType):
+            return fix_generator_encoding(result)
         if isinstance(result, (list, tuple)):
-            return _fix_list_encoding(result)
+            return fix_list_encoding(result)
         if isinstance(result, str):
-            return _to_unicode(result)
+            return to_unicode(result)
 
         return result
 
@@ -88,7 +90,7 @@ def ss(var):
     :return: Converted string
     """
 
-    var = _to_unicode(var)
+    var = to_unicode(var)
 
     try:
         var = var.encode(sickrage.srCore.SYS_ENCODING)
@@ -99,12 +101,15 @@ def ss(var):
             try:
                 var = var.encode(sickrage.srCore.SYS_ENCODING, 'replace')
             except Exception:
-                var = var.encode('utf-8', 'ignore')
+                try:
+                    var = var.encode('utf-8', 'ignore')
+                except Exception:
+                    pass
 
     return var
 
 
-def _fix_list_encoding(var):
+def fix_list_encoding(var):
     """
     Converts each item in a list to Unicode
 
@@ -113,12 +118,29 @@ def _fix_list_encoding(var):
     """
 
     if isinstance(var, (list, tuple)):
-        return filter(lambda x: x is not None, map(_to_unicode, var))
+        var = filter(lambda x: x is not None, map(to_unicode, var))
 
     return var
 
+def fix_generator_encoding(gen):
+    """
+    Converts each item in a list to Unicode
 
-def _to_unicode(var):
+    :param var: List or tuple to convert to Unicode
+    :return: Unicode converted input
+    """
+
+    result = list(gen)
+
+    for i, item in enumerate(result):
+        temp = []
+        for x in item:
+            temp.append(fix_list_encoding(x))
+        result[i] = type(item)(temp)
+
+    return result
+
+def to_unicode(var):
     """
     Converts string to Unicode, using in order: UTF-8, Latin-1, System encoding or finally what chardet wants
 
@@ -184,7 +206,8 @@ def patch_modules():
                 'shutil.move',
                 'shutil.copyfileobj',
                 'shutil.copy',
-                'shutil.copyfile']
+                'shutil.copyfile',
+                'scandir.walk']
 
     def decorate_modules(modules, decorator):
         for module in modules: