+++ sickrage/core/queues/postprocessor.py
@@ -150,18 +150,20 @@ class srPostProcessorQueue(srQueue):
 
             message = logHelper(
                 "An item with directory {} is already being processed in the queue, item updated".format(dirName))
+            return message + "<br\><span class='hidden'>Processing succeeded</span>"
         else:
             item = PostProcessorItem(dirName, nzbName, process_method, force, is_priority, delete_on, failed, proc_type)
             if force_next:
                 with postprocessor_queue_lock:
                     item.run()
                     message = item.result
+                return message
             else:
                 super(srPostProcessorQueue, self).put(item)
                 message = logHelper(
                     "{} post-processing job for {} has been added to the queue".format(proc_type.title(), dirName))
+                return message + "<br\><span class='hidden'>Processing succeeded</span>"
 
-        return message + "<br\><span class='hidden'>Processing succeeded</span>"
 
 
 class PostProcessorItem(srQueueItem):